patch 0.1.0117: Fix GNU/Linux build.
GNU/Linux build was fixed by replacing 'ssize_t' data type with 'int'.
The first one provided by POSIX and can be unavailable on a target
operating system.

+++ include/mcusim/avr/sim/pty.h
@@ -50,15 +50,15 @@ struct MSIM_AVR_PTY {
 
 int MSIM_AVR_PTYOpen(struct MSIM_AVR *mcu);
 int MSIM_AVR_PTYClose(struct MSIM_AVR *mcu);
-ssize_t MSIM_AVR_PTYWrite(struct MSIM_AVR *mcu, uint8_t *buf, uint32_t len);
-ssize_t MSIM_AVR_PTYRead(struct MSIM_AVR *mcu, uint8_t *buf, uint32_t len);
+int MSIM_AVR_PTYWrite(struct MSIM_AVR *mcu, uint8_t *buf, uint32_t len);
+int MSIM_AVR_PTYRead(struct MSIM_AVR *mcu, uint8_t *buf, uint32_t len);
 
 #else
 
-#define MSIM_AVR_PTYOpen(mcu)			0
-#define MSIM_AVR_PTYClose(mcu)			0
-#define MSIM_AVR_PTYWrite(mcu, buf, len)	((ssize_t)-1)
-#define MSIM_AVR_PTYRead(mcu, buf, len)		((ssize_t)-1)
+#define MSIM_AVR_PTYOpen(mcu)			((int)0)
+#define MSIM_AVR_PTYClose(mcu)			((int)0)
+#define MSIM_AVR_PTYWrite(mcu, buf, len)	((int)-1)
+#define MSIM_AVR_PTYRead(mcu, buf, len)		((int)-1)
 
 #endif
